# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ScraptPress API
  description: |
    API profesional para scraping de vehículos de subastas (Copart, IAAI, Manheim).
    
    **Características principales:**
    - Búsqueda síncrona y asíncrona de vehículos
    - Sistema de caché multicapa (Redis + Firestore)
    - Cola asíncrona con Bull Queue para procesamiento en background
    - Rate limiting y protección anti-abuse
    - Validación de seguridad (XSS, SQL Injection, Path Traversal)
    - Request ID tracking para trazabilidad distribuida
    - Métricas de salud y rendimiento en tiempo real
    - Extracción con IA (Gemini Vision) opcional
    - Cumplimiento GDPR completo
    - Server-Sent Events para logs en tiempo real
    
    **Autenticación:**
    - API Key: Header `X-API-Key` para endpoints públicos
    - Admin Token: Header `X-Admin-Token` para operaciones administrativas
    
    **Rate Limiting:**
    - 100 requests/minuto por IP/API Key
    - 10 búsquedas simultáneas por usuario
    - Límites personalizados para API Keys premium
    
  version: 3.1.0
  contact:
    name: ScraptPress Support
    email: support@sumtrading.us
    url: https://github.com/RenePerezRodriguez/ScraptPress
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: https://scrap.sumtrading.us/api
    description: Production Server (Cloud Run)
  - url: http://localhost:3000/api
    description: Local Development Server
...
paths:
...
  /gdpr/access-data:
    post:
      tags:
        - GDPR
      summary: Solicitar acceso a datos personales
      description: |
        Solicitud GDPR Artículo 15 - Derecho de acceso.
        El usuario recibirá sus datos por email en 7 días hábiles.
      operationId: requestDataAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Solicitud recibida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GDPRResponse'
  /gdpr/delete-data:
    post:
      tags:
        - GDPR
      summary: Solicitar eliminación de datos
      description: |
        Solicitud GDPR Artículo 17 - Derecho al olvido.
        Datos eliminados en 30 días según normativa.
      operationId: requestDataDeletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                reason:
                  type: string
      responses:
        '200':
          description: Solicitud recibida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GDPRResponse'
  /gdpr/data-portability:
    post:
      tags:
        - GDPR
      summary: Solicitar portabilidad de datos
      description: |
        Solicitud GDPR Artículo 20 - Derecho a la portabilidad.
        Datos en formato JSON/CSV en 7 días hábiles.
      operationId: requestDataPortability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Solicitud recibida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GDPRResponse'
  /gdpr/consent-record:
    get:
      tags:
        - GDPR
      summary: Obtener registro de consentimientos
      description: Consulta el estado actual de consentimientos del usuario
      operationId: getConsentRecord
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Registro de consentimientos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRecord'
  /gdpr/update-consent:
    post:
      tags:
        - GDPR
      summary: Actualizar consentimientos
      description: Actualiza las preferencias de cookies y procesamiento
      operationId: updateConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - consents
              properties:
                email:
                  type: string
                  format: email
                consents:
                  type: object
                  properties:
                    necessary:
                      type: boolean
                    analytics:
                      type: boolean
                    marketing:
                      type: boolean
      responses:
        '200':
          description: Consentimientos actualizados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
  /gdpr/requests:
    get:
      tags:
        - GDPR
      summary: Ver todas las solicitudes GDPR (Admin)
      description: Lista todas las solicitudes GDPR pendientes
      operationId: getGDPRRequests
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRequests:
                    type: integer
                  byType:
                    type: object
                    properties:
                      access:
                        type: integer
                      delete:
                        type: integer
                      portability:
                        type: integer
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/GDPRRequest'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key para autenticación de endpoints públicos.
        También se acepta en header `Authorization` como `Bearer {key}`.
        
        **Obtención:**
        Contacta a support@sumtrading.us para obtener tu API Key.
        
    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: |
        Token de administrador para operaciones sensibles.
        Solo para personal autorizado.
  schemas:
...
    GDPRResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        details:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    GDPRRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        type:
          type: string
          enum: [access, delete, portability]
        reason:
          type: string
        timestamp:
          type: string
          format: date-time
    ConsentRecord:
      type: object
      properties:
        email:
          type: string
        timestamp:
          type: string
          format: date-time
        consents:
          type: object
          properties:
            necessary:
              type: boolean
            analytics:
              type: boolean
            marketing:
              type: boolean
        rights:
          type: object
          description: Descripción de derechos GDPR
...